requirements.txt
streamlit==1.31.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
flask==3.0.0
pytest==8.0.0
Werkzeug==3.0.1